feat(platform): adding stackblitz support

+++ src/cli/switch-to-ts.ts
@@ -3,9 +3,28 @@
 await global.cli("set-env-var", "KIT_MODE", "ts")
 process.env.KIT_MODE = "ts"
 
-await copyFile(
-  kitPath("templates", "config", "tsconfig.json"),
-  kenvPath("tsconfig.json")
+let tsConfigPath = kenvPath("tsconfig.json")
+let tsConfigExists = await pathExists(tsConfigPath)
+let tsDefaultTemplatePath = kenvPath(
+  "templates",
+  "default.ts"
 )
+let tsDefaultTemplateExists = await pathExists(
+  tsDefaultTemplatePath
+)
+
+if (!tsConfigExists) {
+  await copyFile(
+    kitPath("templates", "config", "tsconfig.json"),
+    tsConfigPath
+  )
+}
+
+if (!tsDefaultTemplateExists) {
+  await copyFile(
+    kitPath("templates", "scripts", "default.ts"),
+    tsDefaultTemplatePath
+  )
+}
 
 export {}
